refactor: extract DeadAnt from AntGroup to render dead ants on back
