+++ src/components/AntGroup/index.tsx
@@ -2,10 +2,11 @@ import { useEffect } from 'react';
 import type { RefObject } from 'react';
 import { Layer } from 'react-konva';
 import type Konva from 'konva';
-import DeadAnt from './DeadAnt';
+import DeadAnt from '../DeadAntGroup/DeadAnt';
 import { useAntKiller } from '@hooks';
 
 import Ant from './Ant';
+import { AntData } from '@recoil';
 
 type AntGroupProps = {
   stageRef: RefObject<Konva.Stage> | null;
@@ -34,9 +35,7 @@ const AntGroup = ({ stageRef }: AntGroupProps) => {
     };
   }, [ants]);
 
-  return (
-    <Layer>{ants.map((ant, idx) => (ant.dead ? <DeadAnt {...ant} key={idx} /> : <Ant {...ant} key={idx} />))}</Layer>
-  );
+  return <Layer>{ants.map((ant, idx) => (ant.dead ? null : <Ant {...ant} key={idx} />))}</Layer>;
 };
 
 export default AntGroup;
