+++ src/components/DeadAntGroup/DeadAnt.tsx
