+++ src/components/index.ts
@@ -5,6 +5,7 @@ export { default as DevBackground } from './DevBackground';
 export { default as Highlighter } from './Highlighter';
 export { default as Stamp } from './Stamp';
 export { default as AntGroup } from './AntGroup';
+export { default as DeadAntGroup } from './DeadAntGroup';
 export { default as Fireworks } from './Fireworks/Fireworks';
 export { default as ContentWrapper } from './ContentWrapper';
 export { default as Hammer } from './Hammer';
